CF-1169

+++ i21Database/dbo/Stored Procedures/uspCFInsertTransactionRecord.sql
@@ -620,7 +620,8 @@ BEGIN
 				SET @intVehicleId =
 				(SELECT TOP 1 intVehicleId
 				FROM @tblCFNumericVehicle
-				WHERE CAST(strVehicleNumber AS INT) = CAST(@strVehicleId AS INT))
+				WHERE CAST(strVehicleNumber AS BIGINT) = CAST(@strVehicleId AS BIGINT))
+
 
 			END
 			ELSE
